重みの型をdoubleにした

+++ source/learner.cpp
@@ -31,7 +31,7 @@ namespace Eval
 
 namespace
 {
-  using WeightType = float;
+  using WeightType = double;
 
   enum WeightKind {
     WEIGHT_KIND_COLOR,
